카테고리 최저가/최고가 업데이트 구현

+++ src/main/kotlin/com/njhyuk/codi/core/product/domian/Product.kt
@@ -15,7 +15,7 @@ data class Product(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     val id: Long = 0L,
-    var price: Double,
+    var price: Long,
     var category: String,
     var brand: String,
     val createdAt: LocalDateTime = LocalDateTime.now(),
